+++ packages/apps/human-app/frontend/src/shared/schemas/index.ts
@@ -1,3 +1,3 @@
 export * from './validate-address-schema';
 export * from './url-domain-schema';
-export * from './sign-in-schema';
+export * from './auth-tokens-schema';
