--- packages/apps/human-app/frontend/src/shared/types/index.ts
@@ -1,1 +1,0 @@
-export * from './sign-in-types';
