--- packages/apps/human-app/frontend/src/shared/types/sign-in-types.ts
@@ -1,9 +1,0 @@
-import { type z } from 'zod';
-import {
-  type signInDtoSchema,
-  type signInSuccessResponseSchema,
-} from '../schemas';
-
-export type SignInDto = z.infer<typeof signInDtoSchema>;
-
-export type SignInSuccessResponse = z.infer<typeof signInSuccessResponseSchema>;
